+++ src/main.tsx
@@ -1,29 +1,11 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import ErrorPage from "./pages/ErrorPage";
-import Cart from "./pages/Cart";
-import App from "./App";
-import Detail from "./pages/Detail";
+import { RouterProvider } from "react-router-dom";
+import router from "./router/rotuer";
 
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <App />,
-    children: [
-      {
-        path: "cart",
-        element: <Cart />,
-      },
-      {
-        path: "detail/:productId",
-        element: <Detail />,
-      },
-    ],
-    errorElement: <ErrorPage />,
-  },
-]);
+
+
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
