+++ src/pages/auth/Join.tsx
@@ -1,0 +1,11 @@
+
+
+const Join = () => {
+  return (
+    <>
+      회원가입 
+    </>
+  )
+}
+
+export default Join
