+++ src/pages/auth/Loigin.tsx
@@ -1,0 +1,11 @@
+
+
+const Login = () => {
+  return (
+    <>
+      로그인 
+    </>
+  )
+}
+
+export default Login
