+++ src/pages/myPage/Mypage.tsx
@@ -1,0 +1,13 @@
+
+
+
+
+const Mypage = () => {
+  return (
+    <>
+      회원가입 
+    </>
+  )
+}
+
+export default Mypage
